- Rearranged navigators so that bottom tab nav doesnt render on top of drawer
