+++ src/components/BottomTabNav.jsx
@@ -33,6 +33,7 @@ export default function BottomTabNav({ state, descriptors, navigation }) {
 
         return (
           <TouchableOpacity
+            key={label}
             accessibilityRole="button"
             accessibilityState={isFocused ? { selected: true } : {}}
             accessibilityLabel={options.tabBarAccessibilityLabel}
